requirements.txt
# --- Web Framework (FastAPI) ---
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# --- AI & Vector Database ---
google-generativeai>=0.3.0
pinecone>=5.0.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2
groq>=0.5.0

# --- Data Processing & Extraction ---
PyPDF2>=3.0.0
pymupdf>=1.23.0
camelot-py[cv]>=0.10.1
opencv-python-headless>=4.8.0
pandas>=2.0.0
Pillow>=10.0.0

# --- Cloud Storage (Azure) ---
azure-storage-blob>=12.17.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0